requirements.txt
kagglehub==0.2.9
pandas==2.1.3
matplotlib==3.8.0
numpy==1.26.2
Pillow==10.2.0
nltk==3.9.1
tqdm==4.67.1
torchinfo==1.8.0